Working Image deconstruct and send What has been done?         - Client can now reduce image to 4096 byte packets         - Server can reconstruct images reduced sent over by client
What needs to be done?
        - Slight image off, could be division offset?
        - Needs slight clean up

+++ Examples/SendImage/Server/src/ImageServer.h
@@ -9,7 +9,7 @@ class ImageServer : public Server
 
                 unsigned char * ImageBuffer;
                 unsigned char   ReceiveBuffer[MaxSize];
-                uint8_t         PacketsToSend; 
+                int             PacketsToSend; 
                 uint32_t        ImageSize;
                 
 	public:
